[binoculars] compute offset0 only once and store it in the cube
this way the add_space method can not be wrong. the offset0 is
necessarely the one from the cube and not by mistake another one.

the burden is to maintain this offset0 each time we change the cube
axes (aka when we merge it with another cube or a space).