[contrib/haskell] replace LenP by ChunkP
This way we can extract for each file the first and the last usable frame.
The idea is to pre-compute the final cube size from the begining in order
to speed (a lot ?) the computation.

We found that 135000 frame of xpad s140 can be read from the ruche and
projected in only 6 minutes. But the complete computation take a lot more
du to the  malloc of Cube of different size.