[haskell/contrib] use the right vector length when setting geometry axes
A geometry expect the right number of axes, the given axes lenght should
be the length of the given axes values and not the geometry expected one.
This way a list with the wrong length could be used for computation.

The flymedV computation was wrong due to the missing beta axis.