[contrib/haskell] compute the number of thread using the number of cores
the number of thread can be set directly in the binary via compile time
+RTS -N<value> -RTS options, or on the command line.
The user can select a maximum number of core from the config file, which
is constrained by the real hardware number of cores.

so set the number of threads with the minimum of the these three
possibilities.